eliminacion de torre

+++ controller/torreController.js
@@ -1,6 +1,7 @@
 const Torre = require("../models/torres");
 const Piso = require("../models/pisos");
 const Beca = require("../models/becas");
+const Cuarto = require("../models/cuartos");
 
 const getTorre = async () => {
   try {
@@ -25,10 +26,10 @@ const getTorre = async () => {
 
 const createTorre = async (jefe_torre, nombre_torre, pisoId) => {
   try {
-    const torre = await Torre.create({ 
-      jefe_torre, 
-      nombre_torre, 
-      pisoId 
+    const torre = await Torre.create({
+      jefe_torre,
+      nombre_torre,
+      pisoId,
     });
     return torre;
   } catch (error) {
@@ -50,9 +51,18 @@ const updateTorre = async (id, jefe_torre, nombre_torre, pisoId) => {
 
 const deleteTorre = async (id) => {
   try {
+    // Verificar si hay cuartos asociados a la torre
+    const cuartosAsociados = await Cuarto.count({ where: { torreid: id } });
+    if (cuartosAsociados > 0) {
+      throw new Error(
+        "No se puede eliminar la torre porque tiene cuartos asociados."
+      );
+    }
+
     const torre = await Torre.destroy({ where: { id } });
     return torre;
   } catch (error) {
+    console.error("Error al eliminar la torre:", error);
     throw error;
   }
 };
